+++ modules/cms/src/admin/resources/js/cmsadmin.js
@@ -172,7 +172,7 @@
 				data : '='
 			},
 			templateUrl : 'createform.html',
-			controller : function($scope, $http, $filter, ServiceMenuData, Slug, ServiceLanguagesData, AdminToastService) {
+			controller : function($scope, $http, $filter, ServiceMenuData, ServiceLanguagesData, AdminToastService) {
 
 				$scope.error = [];
 				$scope.success = false;
@@ -223,12 +223,12 @@
 				});
 
 				$scope.aliasSuggestion = function() {
-					$scope.data.alias = Slug.slugify($scope.data.title);
+					$scope.data.alias = $filter('slugify')($scope.data.title);
 				};
 
 				$scope.$watch('data.alias', function(n, o) {
 					if (n!=o && n!=null) {
-						$scope.data.alias = Slug.slugify(n);
+						$scope.data.alias = $filter('slugify')(n);
 					}
 				});
 
@@ -446,7 +446,7 @@
 		};
 	});
 
-	zaa.controller("CopyPageController", function($scope, $http, AdminToastService, Slug) {
+	zaa.controller("CopyPageController", function($scope, $http, AdminToastService) {
 
 		var headers = {"headers" : { "Content-Type" : "application/x-www-form-urlencoded; charset=UTF-8" }};
 
@@ -480,7 +480,8 @@
 		});
 		
 		$scope.aliasSuggestion = function() {
-			$scope.itemSelection.alias = Slug.slugify($scope.itemSelection.title);
+			
+			$scope.itemSelection.alias = $filter('slugify')($scope.itemSelection.title);
 		};
 
 		$scope.loadItems = function() {
@@ -981,7 +982,7 @@
 	/**
 	 * @param $scope.lang from ng-repeat
 	 */
-	zaa.controller("NavItemController", function($scope, $rootScope, $http, $timeout, Slug, ServiceMenuData, AdminLangService, AdminToastService, ServiceLiveEditMode, ServiceLayoutsData) {
+	zaa.controller("NavItemController", function($scope, $rootScope, $http, $filter, $timeout, ServiceMenuData, AdminLangService, AdminToastService, ServiceLiveEditMode, ServiceLayoutsData) {
 
 		$scope.loaded = false;
 
@@ -1113,7 +1114,7 @@
 
 		$scope.$watch('itemCopy.alias', function(n, o) {
 			if (n!=o && n!=null) {
-				$scope.itemCopy.alias = Slug.slugify(n);
+				$scope.itemCopy.alias = $filter('slugify')(n);
 			}
 		});